pyproject.toml
[project]
name = "git-mcp-server"
version = "0.1.0"
description = "MCP server for local git operations - commit, branch, push, pull, sync"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Rafael Riesco", email = "rriesco@users.noreply.github.com"}
]
keywords = ["mcp", "git", "claude", "anthropic", "automation", "version-control"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.1.0",
    "gitpython>=3.1.0",
    "pydantic>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/rriesco/git-mcp-server"
Repository = "https://github.com/rriesco/git-mcp-server"
Issues = "https://github.com/rriesco/git-mcp-server/issues"

[project.scripts]
git-mcp-server = "git_mcp_server.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Fast unit tests with mocks (run on every commit)",
    "integration: Slower tests with real API calls (run nightly or before release)",
]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--cov=git_mcp_server",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.mypy]
strict = true
warn_unreachable = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
disallow_any_generics = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []
